share shell env bootstrap for PATH and proxy vars
