+++ crates/luban_server/src/lib.rs
@@ -11,6 +11,7 @@ mod mentions;
 mod project_avatars;
 pub mod pty;
 pub mod server;
+pub mod shell_env;
 mod telegram;
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
